Disambiguate and simplify input/output options
Previously, if running as a child process, imagemin would output to
stdout if the parent program tries to capture stdout.  This causes a
problem if the parent program normally writes to stdout itself, but
sometimes gets wrapped.  Further, writing multiple files back to stdout
just doesn't make sense.  Now, imagemin only pipes to stdout if it's
reading from stdin or if exactly one file is provided and no destination
directory is provided.  If a destination directory is provided, it is
always used.

Additionally, inferring directories was sometimes based on whether the
potential directory name had a dot in it, assuming that no directories
have dots in them. Now the destination directory is always assumed to be
a directory, regardless of its name.

One additional input format is allowed if no output (-o) is provided:

 $ imagemin in.png out.png

Finally, the single character command-line flags for optimization
parameters have been removed. '-p' now aliases '--plugin', and '-o' now
aliases '--output'.  This is done to prioritize the parameters that are
always valid, regardless of the plugins used.

update-help